+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## v0.3.3 Beta 3 (WIP)
 
 - Add force_range for TensorrtTensor, borrow_from, active_instances_grp for BaselineSchedule
-- fix TorchScriptTensor for batching input; docs for training-tool
+- fix TorchScriptTensor for batching input; update docs for training-tool
 - test_tools: Add output of median GPU usage rate
+- fix cuda arch error for pytorch 1.13.1 on NVIDIA A10
+- The backend does not allow the deletion of the node_name key, as it would cause the scheduling system to crash.
+- Further support for the Python backend has been added.
 
 
 ## v0.3.3 Beta 2 (2023.11.05)